+++ src/components/icons/PolygonIcon.tsx
@@ -1,0 +1,15 @@
+import type { FC } from 'react'
+
+const PolygonIcon: FC = () => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width="44"
+    height="25"
+    viewBox="0 0 44 25"
+    fill="none"
+  >
+    <path d="M21.6201 0L43.1742 24.2752H0.0658875L21.6201 0Z" fill="white" />
+  </svg>
+)
+
+export default PolygonIcon