+++ src/components/icons/ShareIcon.tsx
@@ -1,0 +1,47 @@
+const ShareIcon = () => (
+  <svg
+    width="28"
+    height="28"
+    viewBox="0 0 28 28"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg"
+  >
+    <path
+      d="M7 17.5C8.933 17.5 10.5 15.933 10.5 14C10.5 12.067 8.933 10.5 7 10.5C5.067 10.5 3.5 12.067 3.5 14C3.5 15.933 5.067 17.5 7 17.5Z"
+      stroke="#FF3452"
+      strokeWidth="1.75"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+    />
+    <path
+      d="M21 10.5C22.933 10.5 24.5 8.933 24.5 7C24.5 5.067 22.933 3.5 21 3.5C19.067 3.5 17.5 5.067 17.5 7C17.5 8.933 19.067 10.5 21 10.5Z"
+      stroke="#FF3452"
+      strokeWidth="1.75"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+    />
+    <path
+      d="M21 24.5C22.933 24.5 24.5 22.933 24.5 21C24.5 19.067 22.933 17.5 21 17.5C19.067 17.5 17.5 19.067 17.5 21C17.5 22.933 19.067 24.5 21 24.5Z"
+      stroke="#FF3452"
+      strokeWidth="1.75"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+    />
+    <path
+      d="M10.15 12.4834L17.85 8.51672"
+      stroke="#FF3452"
+      strokeWidth="1.75"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+    />
+    <path
+      d="M10.15 15.5167L17.85 19.4834"
+      stroke="#FF3452"
+      strokeWidth="1.75"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+    />
+  </svg>
+)
+
+export default ShareIcon