+++ src/components/search/SuggestionProductCard/SuggestionProductCard.tsx
@@ -1,33 +1,15 @@
 import { Card, CardContent, CardImage } from '@faststore/ui'
+import SuggestionImageFallback from 'src/components/icons/SuggestionImageFallback'
 import { Image } from 'src/components/ui/Image'
-import Price from 'src/components/ui/Price'
-import { useFormattedPrice } from 'src/sdk/product/useFormattedPrice'
 
-// TODO: Remove it when integration is complete
-const PRODUCTS = [
-  {
-    price: 46.26,
-    listPrice: 72.06,
-    name: 'Ergonomic Wooden Bacon',
-    image: [
-      {
-        alternateName: 'rerum',
-        url: 'http://storeframework.vtexassets.com/arquivos/ids/167285/ut.jpg?v=637753017045600000',
-      },
-    ],
-  },
-]
+import './suggestion-product-card.scss'
 
-function SuggestionProductCard({
-  // TODO: Add Props interface and define `product` type
-  product = PRODUCTS[0],
-}) {
-  const {
-    name,
-    listPrice,
-    price,
-    image: [img],
-  } = product
+interface SuggestionProductCardProps {
+  product: any
+}
+
+function SuggestionProductCard({ product }: SuggestionProductCardProps) {
+  const { Name: name, ProductImageURL: imageURL } = product
 
   return (
     <Card
@@ -36,32 +18,18 @@ function SuggestionProductCard({
     >
       <CardContent>
         <CardImage>
-          <Image src={img.url} alt={img.alternateName} width={56} height={56} />
+          <Image
+            src={imageURL}
+            alt={name}
+            width={56}
+            height={56}
+            fallbackImage={<SuggestionImageFallback />}
+          />
         </CardImage>
         <div data-suggestion-product-card-summary>
           <p className="text__title-mini" data-suggestion-product-card-title>
             {name}
           </p>
-          <span data-suggestion-product-card-prices>
-            <Price
-              value={listPrice}
-              formatter={useFormattedPrice}
-              testId="list-price"
-              data-value={listPrice}
-              variant="listing"
-              classes="text__legend"
-              SRText="Original price:"
-            />
-            <Price
-              value={price}
-              formatter={useFormattedPrice}
-              testId="price"
-              data-value={price}
-              variant="spot"
-              classes="text__title-mini"
-              SRText="Price:"
-            />
-          </span>
         </div>
       </CardContent>
     </Card>